feat(billing): refactor payment form to add billing address
