+++ apps/deploy-web/src/components/shared/PaymentMethodForm/PaymentMethodForm.tsx
@@ -1,7 +1,7 @@
 "use client";
 import React, { useState } from "react";
-import { Alert, Button, Input } from "@akashnetwork/ui/components";
-import { PaymentElement, useElements, useStripe } from "@stripe/react-stripe-js";
+import { Alert, Button } from "@akashnetwork/ui/components";
+import { AddressElement, PaymentElement, useElements, useStripe } from "@stripe/react-stripe-js";
 
 interface PaymentMethodFormProps {
   onSuccess: () => void;
@@ -19,7 +19,6 @@ export const PaymentMethodForm: React.FC<PaymentMethodFormProps> = ({
   const stripe = useStripe();
   const elements = useElements();
   const [error, setError] = useState<string | null>(null);
-  const [cardholderName, setCardholderName] = useState("");
   const [isProcessing, setIsProcessing] = useState(false);
 
   const handleSubmit = async (e: React.FormEvent) => {
@@ -27,18 +26,13 @@ export const PaymentMethodForm: React.FC<PaymentMethodFormProps> = ({
     if (!stripe || !elements) {
       return;
     }
+
     setError(null);
     setIsProcessing(true);
+
     try {
       const { error: setupError, setupIntent } = await stripe.confirmSetup({
         elements,
-        confirmParams: {
-          payment_method_data: {
-            billing_details: {
-              name: cardholderName
-            }
-          }
-        },
         redirect: "if_required"
       });
 
@@ -58,22 +52,29 @@ export const PaymentMethodForm: React.FC<PaymentMethodFormProps> = ({
   };
 
   return (
-    <form className={`space-y-4 ${className}`} onSubmit={handleSubmit}>
-      <Input
-        id="cardholderName"
-        type="text"
-        value={cardholderName}
-        onChange={e => setCardholderName(e.target.value)}
-        placeholder="Name on card"
-        required
-        label="Cardholder Name"
-      />
-      <PaymentElement />
+    <form className={`space-y-6 ${className}`} onSubmit={handleSubmit}>
+      {/* Billing Address Section */}
+      <div className="space-y-4">
+        <h3 className="text-left text-sm font-semibold text-muted-foreground">Billing Address</h3>
+        <AddressElement
+          options={{
+            mode: "billing"
+          }}
+        />
+      </div>
+
+      {/* Payment Element */}
+      <div className="space-y-2">
+        <h3 className="text-left text-sm font-semibold text-muted-foreground">Card Information</h3>
+        <PaymentElement />
+      </div>
+
       {error && (
         <Alert className="mt-4" variant="destructive">
           {error}
         </Alert>
       )}
+
       <Button type="submit" className="w-full" disabled={isProcessing}>
         {isProcessing ? processingText : buttonText}
       </Button>
